refactor(components): move iconMap to constants directory
